fix: editor_plugin.h was moved to editor/plugins
