+++ editor/fast_noise_2/fast_noise_2_editor_plugin.h
@@ -1,7 +1,7 @@
 #ifndef FAST_NOISE_2_EDITOR_PLUGIN_H
 #define FAST_NOISE_2_EDITOR_PLUGIN_H
 
-#include <editor/editor_plugin.h>
+#include <editor/plugins/editor_plugin.h>
 
 namespace zylann {
